# docs/swagger.yaml
components:
  schemas:
    apperrors.AppError:
      properties:
        code:
          type: string
        message:
          type: string
        params:
          additionalProperties: {}
          type: object
      type: object
    database.GetMailQueueItemsByTaskIdRow:
      properties:
        attempts:
          type: integer
        created_at:
          type: string
        error:
          type: string
        id:
          type: string
        next_attempt_at:
          type: string
        recipient_email:
          type: string
        recipient_full_name:
          type: string
        status:
          $ref: '#/components/schemas/database.NullMailQueueStatus'
      type: object
...
    handlers.MailTaskItem:
      properties:
        audience:
          $ref: '#/components/schemas/handlers.SendAudience'
        body_variables:
          items:
            type: integer
          type: array
          uniqueItems: false
        created_at:
          type: string
        id:
          type: string
        mail_list_id:
          type: string
        mail_list_name:
          type: string
        recipient_counts:
          $ref: '#/components/schemas/handlers.QueueCounts'
        sent_by:
          type: string
        status:
          $ref: '#/components/schemas/handlers.SendStatus'
        template_id:
          type: string
        template_key:
          type: string
        template_name:
          type: string
      type: object
...
    requests.CreateMailTask:
      properties:
        body_variables:
          additionalProperties: {}
          type: object
        mail_list_id:
          type: string
        template_id:
          type: string
      required:
      - mail_list_id
      - template_id
      type: object
    requests.CreateMailingList:
      properties:
        name:
          type: string
      required:
      - name
      type: object
...
    requests.SendSingleMail:
      properties:
        body_variables:
          additionalProperties: {}
          type: object
        recipient_email:
          type: string
        recipient_full_name:
          type: string
        template_id:
          type: string
        template_key:
          type: string
      required:
      - recipient_email
      type: object
...
info:
  contact:
    email: hello@enesgenc.dev
    name: Enes Genç
    url: https://enesgenc.dev
  description: This is the API documentation for Skymail.
  title: Skymail
  version: "1.0"
openapi: 3.1.0
paths:
  /mail_tasks:
    get:
      description: Get a list of all mail tasks with pagination, newest first, each
        with its derived status, recipients by status and audience. The status filter
        keeps only sends with that derived status; X-Total-Count counts the filtered
        sends. Keycloak group names are looked up within 2 seconds for the whole page
        and are null past that.
      parameters:
      - description: Start index
        in: query
        name: _start
        schema:
          type: integer
      - description: End index
        in: query
        name: _end
        schema:
          type: integer
      - description: 'Derived send status: failed (a recipient failed, or none was
          queued a minute after the send), sending (none failed and some pending or
          processing, or none queued yet within that minute), sent (none failed or
          queued, some sent)'
        in: query
        name: status
        schema:
          enum:
          - failed
          - sending
          - sent
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/handlers.MailTaskItem'
                type: array
          description: OK
          headers:
            X-Total-Count:
              description: Number of sends matching the filter
              schema:
                type: integer
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: List all mail tasks
      tags:
      - Mail
    post:
      description: Create a new mail task and queue emails for all recipients. Supports
        both internal mailing lists and Keycloak groups.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.CreateMailTask'
                description: Mail task details
                summary: task
        description: Mail task details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Create a new mail task
      tags:
      - Mail
  /mail_tasks/{id}:
    get:
      description: Get details of a specific mail task by its ID, with its derived
        status, recipients by status and audience, as the send list shows it.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.MailTaskItem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Get a mail task by ID
      tags:
      - Mail
  /mail_tasks/{id}/queue:
    get:
      description: Get a send's recipients, one queue row each, newest first with
        the row id breaking ties so pages neither repeat nor skip a recipient. The
        status filter keeps only recipients in that queue status; X-Total-Count counts
        the filtered recipients.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Start index
        in: query
        name: _start
        schema:
          type: integer
      - description: End index
        in: query
        name: _end
        schema:
          type: integer
      - description: Recipient queue status
        in: query
        name: status
        schema:
          enum:
          - pending
          - processing
          - sent
          - failed
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/database.GetMailQueueItemsByTaskIdRow'
                type: array
          description: OK
          headers:
            X-Total-Count:
              description: Number of recipients matching the filter
              schema:
                type: integer
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Get mail queue items for a task
      tags:
      - Mail
  /mail_tasks/single:
    post:
      description: Send a single email to a specific recipient using a template.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.SendSingleMail'
                description: Mail details
                summary: mail
        description: Mail details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Send a single email
      tags:
      - Mail
  /mail_tasks/summary:
    get:
      description: 'Queue rows (one per recipient) by status; sends by derived status,
        each equal to the X-Total-Count of the send list filtered by it; mail sent
        per Europe/Istanbul day over the last days (zero-filled, oldest first, ending
        today); and the latest sends. Derived status: failed (a recipient failed,
        or none was queued a minute after the send), sending (none failed and some
        pending or processing, or none queued yet within that minute), sent (none
        failed or queued, some sent).'
      parameters:
      - description: Days in the daily series, ending today in Europe/Istanbul (1-90)
        in: query
        name: days
        schema:
          default: 30
          type: integer
      - description: Latest sends to list, newest first (1-20)
        in: query
        name: recent
        schema:
          default: 5
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.SendSummary'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Summarise mail sends
      tags:
      - Mail
  /mailing_lists:
    get:
      description: Get a list of all mailing lists (internal + Keycloak groups) with
        pagination.
      parameters:
      - description: Start index
        in: query
        name: _start
        schema:
          type: integer
      - description: End index
        in: query
        name: _end
        schema:
          type: integer
      - description: 'Lifecycle filter: current, inactive, all'
        in: query
        name: lifecycle
        schema:
          default: current
          enum:
          - current
          - inactive
          - all
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/handlers.MailingListItem'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: List all mailing lists
      tags:
      - Lists
    post:
      description: Create a new mailing list with the provided name.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.CreateMailingList'
                description: List details
                summary: list
        description: List details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/database.MailingList'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Create a new mailing list
      tags:
      - Lists
  /mailing_lists/{id}:
    delete:
      description: Archive a mailing list without removing recipients or historical
        mail tasks. Repeating the request is safe.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Archive a mailing list
      tags:
      - Lists
    get:
      description: Get details of a specific mailing list by its ID. Falls back to
        Keycloak groups.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.MailingListItem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Get a mailing list by ID
      tags:
      - Lists
    patch:
      description: Update an existing mailing list with the provided ID and details.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.UpdateMailingList'
                description: List details
                summary: list
        description: List details
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/database.MailingList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Update a mailing list
      tags:
      - Lists
  /mailing_lists/{id}/recipients:
    get:
      description: Get all recipients in a mailing list or Keycloak group with pagination.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Start index
        in: query
        name: _start
        schema:
          type: integer
      - description: End index
        in: query
        name: _end
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/database.Recipient'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: List recipients of a mailing list
      tags:
      - Lists
    post:
      description: Add a new recipient or link an existing one by email to a mailing
        list.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.AddRecipient'
                description: Recipient details
                summary: recipient
        description: Recipient details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/database.Recipient'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Add a recipient to a mailing list
      tags:
      - Lists
  /mailing_lists/{id}/recipients/{recipientId}:
    delete:
      description: Remove a specific recipient from a mailing list by their IDs.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Recipient ID
        in: path
        name: recipientId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Remove a recipient from a mailing list
      tags:
      - Lists
  /mailing_lists/{id}/restore:
    post:
      description: Restore an archived internal mailing list. Repeating the request
        is safe.
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.MailingListItem'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Conflict
      summary: Restore an archived mailing list
      tags:
      - Lists
...
  /templates/{id}/restore:
    post:
      description: Restore an archived email template. Repeating the request is safe.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Conflict
      summary: Restore an archived email template
      tags:
      - Templates
...
servers:
- url: skymail-api.yildizskylab.com/v1
...